requirements.txt
boto3
scanpy
anndata
scvelo
python-igraph
louvain
pandasql
black
flake8
pytest
pytest-cov
moto
pytest-mock
diffxpy
python-dateutil